+++ src/heading.ts
@@ -14,7 +14,7 @@ import type {LintError} from '../index';
  */
 export class HeadingToken extends sol(fixed(Token)) {
 	override readonly type = 'heading';
-	declare name: string;
+	#level;
 
 	declare childNodes: [Token, SyntaxToken];
 	// @ts-expect-error abstract method
@@ -35,7 +35,7 @@ export class HeadingToken extends sol(fixed(Token)) {
 
 	/** 标题层级 */
 	get level(): number {
-		return Number(this.name);
+		return this.#level;
 	}
 
 	/* NOT FOR BROWSER */
@@ -66,7 +66,7 @@ export class HeadingToken extends sol(fixed(Token)) {
 	 */
 	constructor(level: number, input: [string?, string?], config = Parser.getConfig(), accum: Token[] = []) {
 		super(undefined, config, accum);
-		this.setAttribute('name', String(level));
+		this.#level = level;
 		const token = new Token(input[0], config, accum);
 		token.type = 'heading-title';
 		token.setAttribute('stage', 2);
@@ -109,7 +109,7 @@ export class HeadingToken extends sol(fixed(Token)) {
 		const errors = super.lint(start),
 			innerStr = String(this.firstChild);
 		let refError: LintError | undefined;
-		if (this.name === '1') {
+		if (this.level === 1) {
 			refError = generateForSelf(this, {start}, '<h1>');
 			errors.push(refError);
 		}
@@ -148,7 +148,7 @@ export class HeadingToken extends sol(fixed(Token)) {
 	 * @param n 标题层级
 	 */
 	setLevel(n: number): void {
-		this.setAttribute('name', String(Math.min(Math.max(n, 1), 6)));
+		this.#level = Math.min(Math.max(n, 1), 6);
 	}
 
 	/** 移除标题后的不可见内容 */